docker/grafana/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
grafanalib==0.6.3
idna==3.3
noroutine-grit==0.0.4
pydantic==1.9.0
pydantic-argparse==0.5.0
python-decouple==3.6
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
typing_extensions==4.2.0
urllib3==1.26.9
